+++ packages/ui/src/components/input-otp/input-otp.stories.tsx
@@ -22,9 +22,6 @@ export const Default: Story = {
         <InputOTPSlot index={0} />
         <InputOTPSlot index={1} />
         <InputOTPSlot index={2} />
-      </InputOTPGroup>
-      <InputOTPSeparator />
-      <InputOTPGroup>
         <InputOTPSlot index={3} />
         <InputOTPSlot index={4} />
         <InputOTPSlot index={5} />
@@ -105,6 +102,24 @@ export const SizeXl: Story = {
   ),
 }
 
+export const Connected: Story = {
+  render: () => (
+    <InputOTP maxLength={6}>
+      <InputOTPGroup size="connected">
+        <InputOTPSlot index={0} />
+        <InputOTPSlot index={1} />
+        <InputOTPSlot index={2} />
+      </InputOTPGroup>
+      <InputOTPSeparator />
+      <InputOTPGroup size="connected">
+        <InputOTPSlot index={3} />
+        <InputOTPSlot index={4} />
+        <InputOTPSlot index={5} />
+      </InputOTPGroup>
+    </InputOTP>
+  ),
+}
+
 export const Disabled: Story = {
   render: () => (
     <InputOTP maxLength={6} disabled>
@@ -112,9 +127,6 @@ export const Disabled: Story = {
         <InputOTPSlot index={0} />
         <InputOTPSlot index={1} />
         <InputOTPSlot index={2} />
-      </InputOTPGroup>
-      <InputOTPSeparator />
-      <InputOTPGroup>
         <InputOTPSlot index={3} />
         <InputOTPSlot index={4} />
         <InputOTPSlot index={5} />